# docs/api/openapi.yaml
openapi: '3.1.0'
info:
  title: Orchestrate OS API
  version: 2.5.0-draft
  description: 'Governance-only semantic control plane API. Resource-style endpoints for multi-user Postgres-backed persistence
    with append-only audit logging.

    '
  contact:
    name: Orchestrate OS Team
servers:
- url: /api/v2.5
  description: Canonical API base
paths:
...
  /documents/{doc_id}/anchors:
    post:
      operationId: createAnchor
      summary: Create a text anchor on a document
      description: 'Creates a text anchor with fingerprint-based dedup. If an anchor with the same fingerprint already exists,
        returns the existing anchor (200) instead of creating a duplicate. Emits anchor.created audit event. Feature-gated
        behind EVIDENCE_INSPECTOR_V251.

        '
      tags:
      - Evidence Inspector
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnchorCreate'
      responses:
        '201':
          description: Anchor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnchorResponse'
        '200':
          description: Existing anchor returned (fingerprint dedup)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnchorResponse'
        '404':
          description: Document not found or feature disabled
    get:
      operationId: listAnchors
      summary: List anchors for a document
      description: 'Returns paginated list of anchors for a document. Feature-gated behind EVIDENCE_INSPECTOR_V251.

        '
      tags:
      - Evidence Inspector
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/CursorParam'
      - $ref: '#/components/parameters/LimitParam'
      - name: include_deleted
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Anchor list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnchorCollection'
        '404':
          description: Document not found or feature disabled
...
components:
  parameters:
...
    CursorParam:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    CollectionMeta:
      type: object
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
        pagination:
          type: object
          properties:
            cursor:
              type: string
            has_more:
              type: boolean
            limit:
              type: integer
    Anchor:
      type: object
      properties:
        id:
          type: string
          description: ULID with anc_ prefix
        document_id:
          type: string
        workspace_id:
          type: string
        anchor_fingerprint:
          type: string
          description: SHA-256 fingerprint for dedup
        node_id:
          type: string
        char_start:
          type: integer
        char_end:
          type: integer
        selected_text:
          type: string
        field_id:
          type: string
        field_key:
          type: string
        page_number:
          type: integer
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
        version:
          type: integer
        metadata:
          type: object
        selected_text_hash:
          type: string
          description: SHA-256 hash of selected_text, computed server-side
    AnchorCreate:
      type: object
      properties:
        node_id:
          type: string
        char_start:
          type: integer
        char_end:
          type: integer
        selected_text:
          type: string
        field_id:
          type: string
        field_key:
          type: string
        page_number:
          type: integer
        metadata:
          type: object
    AnchorResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Anchor'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    AnchorCollection:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Anchor'
        meta:
          $ref: '#/components/schemas/CollectionMeta'
...
  securitySchemes:
    GoogleOAuth:
      type: oauth2
      description: 'Google OAuth 2.0 (OIDC) for production human users. Required for human-governed endpoints (patch transitions,
        approvals, annotations).

        '
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            email: Read user email
            profile: Read user profile
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Server-issued JWT (1h expiry) obtained after Google OAuth code exchange. Used in Authorization header
        for all authenticated requests.

        '
    ScopedApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'Workspace-scoped API key for sandbox and service ingestion endpoints. Keys carry explicit scopes (e.g.,
        signals:write, batches:write, read:all). Required for service ingestion endpoints (POST /signals, POST /triage-items).
        Accepted as alternative auth on read endpoints.

        '
security:
- BearerToken: []
- ScopedApiKey: []